Flatten WaitlistRow drag/drop behaviours
