+++ src/components/WaitList/ModRow.css
@@ -32,3 +32,21 @@
 .WaitlistRow-remove {
   cursor: pointer;
 }
+
+.WaitlistRow--dropAbove:before,
+.WaitlistRow--dropBelow:after {
+  content: ' ';
+  display: block;
+  background: white;
+  height: 3px;
+  width: 100%;
+  position: absolute;
+}
+
+.WaitlistRow--dropAbove:before {
+  top: -1px;
+}
+
+.WaitlistRow--dropBelow:after {
+  top: 38px;
+}
