+++ src/components/WaitList/ModRow.js
@@ -1,29 +1,66 @@
+import cx from 'clsx';
 import React from 'react';
 import PropTypes from 'prop-types';
-import { useDrop } from 'react-dnd';
+import { useDrag, useDrop } from 'react-dnd';
+import DragIcon from '@material-ui/icons/DragHandle';
+import RemoveIcon from '@material-ui/icons/Close';
 import { WAITLIST_USER } from '../../constants/DDItemTypes';
 import isDraggingNearTopOfRow from '../../utils/isDraggingNearTopOfRow';
-
-import ModRowBase from './ModRowBase';
+import useUserCard from '../../hooks/useUserCard';
+import Avatar from '../Avatar';
+import Username from '../Username';
+import Position from './Position';
 
 const {
+  useCallback,
   useEffect,
   useRef,
   useState,
 } = React;
 
-function ModRow({ style, ...props }) {
-  const { position } = props;
+/**
+ * A Draggable waitlist user row with moderation tools.
+ */
+function ModRow({
+  className,
+  position,
+  user,
+  onMoveUser,
+  onRemoveUser,
+}) {
+  const userCard = useUserCard(user);
+  const onOpenCard = useCallback((event) => {
+    event.preventDefault();
+    userCard.open();
+    // The `userCard.open` reference never changes.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
+  // Drag-drop interactions
+  const handleRef = useRef();
   const [insertAbove, setInsertAbove] = useState(false);
-  const refRow = useRef(null);
+  const [{ isDragging }, connectDragSource, connectDragPreview] = useDrag({
+    type: WAITLIST_USER,
+    item() {
+      return { user };
+    },
+    end(item, monitor) {
+      const result = monitor.getDropResult();
+      if (item.user && result) {
+        onMoveUser(result.position);
+      }
+    },
+    collect(monitor) {
+      return { isDragging: monitor.isDragging() };
+    },
+  });
   const [{ isOver }, connectDropTarget] = useDrop(() => ({
     accept: WAITLIST_USER,
     hover(item, monitor) {
-      setInsertAbove(isDraggingNearTopOfRow(monitor, refRow.current));
+      setInsertAbove(isDraggingNearTopOfRow(monitor, userCard.refAnchor.current));
     },
     drop(item, monitor) {
-      const insertAfter = !isDraggingNearTopOfRow(monitor, refRow.current);
+      const insertAfter = !isDraggingNearTopOfRow(monitor, userCard.refAnchor.current);
       return {
         position: insertAfter ? position + 1 : position,
       };
@@ -34,25 +71,63 @@ function ModRow({ style, ...props }) {
   }), [position]);
 
   useEffect(() => {
-    connectDropTarget(refRow.current);
+    if (userCard.refAnchor.current) {
+      connectDragPreview(userCard.refAnchor.current);
+    }
   });
 
-  const dropIndicator = <div className="WaitlistRow-drop-indicator" />;
+  connectDropTarget(userCard.refAnchor);
+  connectDragSource(handleRef);
 
-  // Need a wrapper div so hovering over the drop indicator
-  // does not change the hover state.
+  const rowClassName = cx(className, {
+    UserRow: true,
+    WaitlistRow: true,
+    'WaitlistRow--moderate': true,
+    'WaitlistRow--dropAbove': isOver && insertAbove,
+    'WaitlistRow--dropBelow': isOver && !insertAbove,
+    'is-dragging': isDragging,
+  });
+
   return (
-    <div style={style} ref={refRow}>
-      {isOver && insertAbove && dropIndicator}
-      <ModRowBase {...props} />
-      {isOver && !insertAbove && dropIndicator}
+    <div
+      className={rowClassName}
+      ref={userCard.refAnchor}
+    >
+      {userCard.card}
+      <Position position={position + 1} />
+      <button
+        type="button"
+        className="WaitlistRow-card"
+        onClick={onOpenCard}
+      >
+        <Avatar
+          className="UserRow-avatar"
+          user={user}
+        />
+        <Username className="UserRow-username" user={user} />
+      </button>
+      <div className="WaitlistRow-tools">
+        <div ref={handleRef} className="WaitlistRow-tool WaitlistRow-handle">
+          <DragIcon />
+        </div>
+        <button
+          type="button"
+          className="WaitlistRow-tool WaitlistRow-remove"
+          onClick={onRemoveUser}
+        >
+          <RemoveIcon />
+        </button>
+      </div>
     </div>
   );
 }
 
 ModRow.propTypes = {
-  style: PropTypes.object, // for react-window
+  className: PropTypes.string,
   position: PropTypes.number.isRequired,
+  user: PropTypes.object.isRequired,
+  onMoveUser: PropTypes.func.isRequired,
+  onRemoveUser: PropTypes.func.isRequired,
 };
 
 export default ModRow;
